Bugfix for an error, when daily backup fails to find upgrade.dat file.; Granger will no longer clear parent names on smilexamines, refer to this post for explanation: http://forum.wurmonline.com/index.php?/topic/68031-wurm-assistant-enrich-your-wurm-experience/&do=findComment&comment=1581931

+++ src/Apps/WurmAssistant/WurmAssistant3/Systems/DataBackups/BackupManager.cs
@@ -58,23 +58,35 @@ namespace AldursLab.WurmAssistant3.Systems.DataBackups
 
             {
                 var upgradeVersionFile = Path.Combine(dataDirectory.DirectoryPath, "upgrade.dat");
-                var backupVersionFile = Path.Combine(currentBackupRootPath, "upgrade.dat");
-                File.Copy(upgradeVersionFile, backupVersionFile);
+                if (File.Exists(upgradeVersionFile))
+                {
+                    var backupVersionFile = Path.Combine(currentBackupRootPath, "upgrade.dat");
+                    File.Copy(upgradeVersionFile, backupVersionFile);
+                }
             }
             {
                 var dataDir = Path.Combine(dataDirectory.DirectoryPath, "Data");
-                var backupDir = Path.Combine(currentBackupRootPath, "Data");
-                DirectoryOps.CopyRecursively(dataDir, backupDir);
+                if (Directory.Exists(dataDir))
+                {
+                    var backupDir = Path.Combine(currentBackupRootPath, "Data");
+                    DirectoryOps.CopyRecursively(dataDir, backupDir);
+                }
             }
             {
                 var dataV2Dir = Path.Combine(dataDirectory.DirectoryPath, "DataV2");
-                var backupV2Dir = Path.Combine(currentBackupRootPath, "DataV2");
-                DirectoryOps.CopyRecursively(dataV2Dir, backupV2Dir);
+                if (Directory.Exists(dataV2Dir))
+                {
+                    var backupV2Dir = Path.Combine(currentBackupRootPath, "DataV2");
+                    DirectoryOps.CopyRecursively(dataV2Dir, backupV2Dir);
+                }
             }
             {
                 var soundBankDir = Path.Combine(dataDirectory.DirectoryPath, "SoundBank");
-                var backupSoundBankDir = Path.Combine(currentBackupRootPath, "SoundBank");
-                DirectoryOps.CopyRecursively(soundBankDir, backupSoundBankDir);
+                if (Directory.Exists(soundBankDir))
+                {
+                    var backupSoundBankDir = Path.Combine(currentBackupRootPath, "SoundBank");
+                    DirectoryOps.CopyRecursively(soundBankDir, backupSoundBankDir);
+                }
             }
 
             var confirmFile = Path.Combine(currentBackupRootPath, "completed.dat");
